Refactor SBSL example, improve physics module separation and code quality
